refactor: tracking location without useEffect and code spliting for Client
