+++ client/src/feature/challenge/previousChallengesSlice.ts
@@ -1,20 +1,38 @@
-// import { RootState } from './../store/store';
 // import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 // // api 요청 함수 import해오기
 
-// interface PreviousChallenges {
-//   value: string
+// interface TargetChallenge {
+//   day: number
+//   contents: string
+//   time: string
+// }
+// interface Challenges {
+//   index: string
+//   success: boolean
+//   type: string | 'MORNING' | 'STUDY' | 'THANKS'
+//   targetDate: number
+//   targetTime: number
+//   histories: TargetChallenge[]
 // }
 
-// const initialState: PreviousChallenges = {
-//   value: 'hi'
+// interface PreviousChallenges {
+//   totalDate: number
+//   totalChall: number
+//   challenges: Challenges[]
+// }
+
+// const initialTotalChallenges: PreviousChallenges = {
+//   totalDate: 0,
+//   totalChall: 0,
+//   challenges: []
 // };
 
 // export const fetchPreviousChallenges = createAsyncThunk(
 //   'mypage/fetchPreviousChallenges',
 //   async (totalData: [], thunkAPI) => {
-//     const response = await 'import해온 API'.fetchPreviousChallenges(value);
+//     const response = await fetchPreviousChallenges(value);
 //     return response.data;
 //   }
 // );
+
 export {};
