--- client/src/lib/constants.ts
@@ -1,1 +1,0 @@
-export {};
