+++ client/src/views/components/UI/Layout.style.ts
@@ -25,6 +25,7 @@ const ScrollContainer = styled(FlexColumn)`
 `;
 
 const ChallengeWrapper = styled.div`
+  word-break: keep-all;
   width: 75%;
   margin-bottom: 15px;
   border-bottom: 1px solid var(--graph-bg-green);
