+++ client/src/views/components/common/MyPageNav.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TapHome, TapHomeBtn } from 'views/pages/Mypage/mypage.style';
+import { TapHomeBtn } from 'views/pages/Mypage/mypage.style';
 import { TapHomeIcon } from 'views/components/UI/atoms/icon.style';
 import { useNavigate } from 'react-router-dom';
 
@@ -7,16 +7,10 @@ const MyPageNav = () => {
   const navigate = useNavigate();
 
   return (
-  <>
-    {
-      location.pathname === '/mypage'
-        ? <TapHome>마이페이지</TapHome>
-        : <TapHomeBtn as='button' onClick={() => navigate('/mypage')}>
-            <TapHomeIcon className="material-symbols-outlined">keyboard_backspace</TapHomeIcon>
-            마이페이지
-          </TapHomeBtn>
-    }
-  </>
+    <TapHomeBtn as='button' onClick={() => navigate('/mypage')}>
+      <TapHomeIcon className="material-symbols-outlined">keyboard_backspace</TapHomeIcon>
+      마이페이지
+    </TapHomeBtn>
   );
 };
