+++ client/src/views/pages/PreviousChallenges/Morning.tsx
@@ -4,12 +4,12 @@ import { TitleIcon } from 'views/components/UI/atoms/icon.style';
 import { Title, Content } from 'views/components/UI/molecules/text.style';
 
 const PreviousMorning = () => {
-  const dummy = [{ day: '2022년 5월 5일', time: '0시 0분 0초' },
-    { day: '2022년 5월 5일', time: '0시 0분 0초' },
-    { day: '2022년 5월 5일', time: '0시 0분 0초' },
-    { day: '2022년 5월 5일', time: '0시 0분 0초' },
-    { day: '2022년 5월 5일', time: '0시 0분 0초' },
-    { day: '2022년 5월 5일', time: '0시 0분 0초' }];
+  const dummy = [{ day: '2022년 5월 5일', time: '2' },
+    { day: '2022년 5월 5일', time: '12' },
+    { day: '2022년 5월 5일', time: '2' },
+    { day: '2022년 5월 5일', time: '2' },
+    { day: '2022년 5월 5일', time: '2' },
+    { day: '2022년 5월 5일', time: '2' }];
 
   return (
     <>
@@ -19,7 +19,8 @@ const PreviousMorning = () => {
           <Title.Day>{el.day}</Title.Day>
           <Content.Diary>
             <TitleIcon className="material-symbols-outlined">wb_sunny</TitleIcon>
-            기상 시간: {el.time}
+            목표: {el.time}시
+            <div>00시 00분 0에 일어났어요</div>
           </Content.Diary>
         </Layout.ChallengeWrapper>
       );
