--- client/src/views/pages/SignupPage/signup.style.ts
@@ -1,20 +1,0 @@
-import styled from 'styled-components';
-
-export const NoticeBox = styled.button`
-  width: fit-content;
-  background: none;
-  padding: 0;
-  margin-bottom: 20px;
-  text-align: left;
-  color: var( --intro-form-darkgreen);
-`;
-
-export const ErrorMessage = styled.div`
-  text-align: center;
-  color: red;
-  margin: 10px;
-`;
-
-// const StyledInput = styled(AuthInput)`
-//   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.05);
-// `;
